Custom Assertion Messages
Adds the possibility to add custom assertion messages when using assert.
This allows to give better feedback when running a test suite. A simple
example for that might be the following:

  test "failed custom assertion" do
    assert(string.empty?, "not empty")
  end

This gives better feedback, because it gives a reason for the failure.

It also makes the step to a custom assertion very easy when we find that
we need this kind of assertion multiple times:

  def assert_empty(string)
    assert(string.empty?, "not empty")
  end

  test "failed custom assertion" do
    assert_empty "foo"
  end