Return the expected error message for a failed assert_raise
Since commit f6b140c assert_raise started using the second argument of assert
for error messages. The error message was always "got nil instead". The reason
is that the local variable "exception" only has a value in case of success. In
case of failure its value is nil.

This commit adds tests to catch the error and fixes assert_raise to construct
the expected error message in case of failure.